Rearranged ZScore in Quant Trading System to better pull dates from today backwards - instead of selecting dates and running forwards to today.
Used ggplot2 to build a 'returns' visualizer for individual securities.

correlation_heatmaps built to add instruments to and measure their correlations as aheat map in ggplot2

summarystats.R is to display a histogram and summary statistics for a given instrument

Changed the name of the "dailyReturnsVisual" to "BETA_capm" as a reminder to further build on some CAPM tools.